Merge branch 'lock-refactoring'
The conflicts were mainly at the intersection of the refactored
consistentkey locking code and parallel testing in Cassandra.
LocalLockMediators were not in keyspace-specific namespaces to keep
parallel Cassandra tests from interfering and causing false local lock
contention.  I think I resolved it in this commit, but if I did make
an error merging, it's probably got something to do with
LocalLockMediator configuration in the new ConsistentKey- and/or
AstyanaxRecipeLocker.

The complete Titan test suite did pass on my box the three consecutive
times that I ran it before committing.

Conflicts:
	doc
	titan-cassandra/src/test/java/com/thinkaurelius/titan/CassandraStorageSetup.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/locking/LocalLockMediators.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/locking/consistentkey/ConsistentKeyLockConfiguration.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/locking/consistentkey/ConsistentKeyLockStore.java
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/locking/consistentkey/ConsistentKeyLockTransaction.java
	titan-test/src/main/java/com/thinkaurelius/titan/diskstorage/LockKeyColumnValueStoreTest.java