Initial Sonar support in pom.xml files
This changeset generates a single JaCoCo coverage file containing the
aggregated coverage data for all Maven modules containing tests.  It
also adds a Sonar plugin configuration that uploads the coverage data
to a Sonar server.  The approach is inelegant, but it seemed to be the
least worst compromise between various tooling limitations.  In
particular, Sonar operates on a per-module basis, and it appears to
discard coverage data for source files outside the module it's
currently considering.  This is a problem for our project, since most
of our tests exercise code across Maven module boundaries.