Refactor poms to allow JDK cross-compiling
The current Titan source can't build on JDK 6 (tested Oracle 6u45)
because TitanVertexQuery inherits two unrelated return types for
the method limit(int); this is supported in 7 but not 6.

Here's the exact wording of the compiler error on Oracle 6u45:

[ERROR] /home/dalaro/tinkerelius/titan/titan-core/src/main/java/com/thinkaurelius/titan/core/TitanVertexQuery.java:[16,7] types com.tinkerpop.blueprints.VertexQuery and com.thinkaurelius.titan.core.BaseVertexQuery are incompatible; both define limit(int), but with unrelated return types

Related bug: http://bugs.sun.com/view_bug.do?bug_id=6908259

However, we can still compile Titan using JDK 7 with -source 1.6
-target 1.6 and -Xbootclasspath:/path/to/jdk/rt.jar.  This does not
throw the error above.

This commit records the first complete set of project poms that can
build for both jars and assemblies -target 1.6 and -target 1.7.  It's
still rough and could use some cleanup.  In particular, the "jre6" and
"standard" profiles that I introduced in this commit are unnecessary
whenever the set of dependencies in each case is identical, differing
only in the dependency classifiers on titan artifacts.  I didn't
realize that at the start, so this commit includes a combination of
approaches for managing dependencies under the two compilation
targets.  Also, the names "jre6" and "standard" for compilation
targeting profiles are bad.