Extending ISPN configuration options
InfinispanCacheStoreManager gains two configuration options:

* conffile (GraphDatabaseConfiguration.STORAGE_CONF_FILE_KEY)
* cacheprefix (ICSM.INFINISPAN_CACHE_NAME_PREFIX_KEY)

The former defaults to null.  When it is non-null, it is interpreted
as a path to an Infinispan XML config file and that file is used to
instantiate the EmbeddedCacheManager instance.  When it is null, the
default Infinispan global configuration takes effect.

The latter defaults to titan.  ICSM now uses cachenames in the form
<prefix>:<dbname>.  For example, under the default configuration, the
cache names are:

* titan:titan_ids
* titan:edgestore
* titan:edgestore_lock_
* titan:vertexindex
* titan:vertexindex_lock_
* titan:edgeindex

The prefix is omitted from the value returned by
InfinispanCacheStore#getName() and intended to be hidden from the
graphdb layer above it.

Why add this feature?  Potential use cases for the configurable prefix
include:

* running multiple distinct ISPN-backed databases concurrently, each
  with its own prefix

* resolving potential cache name conflicts with other ISPN clients
  (i.e. playing nice in a cotenant ISPN deployment)

* switching between different cache configs/tunings without
  necessarily rewriting infinispan.xml every time