Adding StaticBuffer serializer for Infinispan
This should probably be part of the default Infinispan global
configuration if it works, but as of this commit it can be enabled in
infinispan.xml:

<infinispan>
  <global>
    <serialization>
      <advancedExternalizers>
        <advancedExternalizer externalizerClass="com.thinkaurelius.titan.diskstorage.infinispan.ext.StaticBufferAE"/>
      </advancedExternalizers>
    </serialization>
  </global>
</infinispan>

The serializer performs no compression.  It just reads/writes integer
lengths followed by bytes.  I've only tested it on some tiny graph
traversals on EC2 so far, and it worked fine there.