Initial commit of Infinispan transaction support
This introduces Infinispan transaction support backed by the JBoss
Transaction Manager, also known as Narayana and the core module of
which is known as Arjuna.  The implementation uses javax.transaction
where possible, which should theoretically ease a potential switch to
an alternative JTA/JTS transaction provider (say, Atomikos).

This commit also changes the Infinispan store configuration from
non-transactional to transactional for all tests.  I'll add the
nontransactional configuration back in as a separate test execution in
a future commit, so that each test for Infinispan is executed in both
a transactional and nontransactional configuration.

I also removed the Infinispan LockKeyColumnValueStoreTest, since its
semantics make no sense for a store that has builtin key-level locking
that doesn't implement acquireLock.  It may make sense to re-add it
with a nontransactional config.

This code is new, just barely tested, and should be assumed buggy.