Refactor Infinispan persistence configuration
Configuration of SingleFileCacheStore in Infinispan now uses the
storage.directory configuration key.  This key is now optional.  If
not present, then persistence is disable.  If present, then a
SingleFileCacheStore is created on the specified directory path.
Infinispan then reads/writes one file per named cache inside the
specified directory.

I've tested this in bin/gremlin.sh by loadind and persisting trivial
example data, but that's it.